49 remove auth guard from library

+++ apps/api/src/app/user-exercise/user-exercise.service.ts
@@ -3,6 +3,8 @@ import { Workout } from '../../../../../libs/interfaces/workout';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserExerciseEntity } from './user-exercise.entity';
 import { Repository } from 'typeorm';
+import { ErrorCode } from '../../../../../libs/error-code/error-code';
+import { CustomBadRequestException } from '../exceptions/CustomBadRequestException';
 
 @Injectable()
 export class UserExerciseService {
@@ -14,20 +16,24 @@ export class UserExerciseService {
     }
 
     async create(workout: Workout) {
-        const existingWorkout = await this.workoutRepository.findOne({
-            where: {
-                user: {
-                  id: workout.user.id
-                },
-                exercise: {
-                    id: workout.exercise.id
+        if (!workout?.user?.id || !workout?.exercise?.id) {
+            throw new CustomBadRequestException(ErrorCode.userMustBeLoggedToContinue, 'Vous devez être connecté pour continuer');
+        } else {
+            const existingWorkout = await this.workoutRepository.findOne({
+                where: {
+                    user: {
+                        id: workout.user.id
+                    },
+                    exercise: {
+                        id: workout.exercise.id
+                    }
                 }
+            });
+            if (existingWorkout) {
+                return await this.workoutRepository.remove(existingWorkout);
             }
-        });
-        if (existingWorkout) {
-            return this.workoutRepository.remove(existingWorkout);
-        }
 
-        return this.workoutRepository.save(workout);
+            return await this.workoutRepository.save(workout);
+        }
     }
 }
